fix PR 3
